Generate `Torrent` and `TorrentField` using procedural macro.
Introduced procedural macro called `torrent_proc` which lives in
it's own crate in `torrent_macro/`. The macro generates both the
struct `Torrent` and the enum `TorrentField` (which replaces the
`Fields` enum) from single definition of the fields.

This wasn't propably a smart move, as now we have even more code
than we started with, in two separate crates and heavily using
very unstable internal compiler apis. This is going to break propbaly
a lot. :/ It's cool though.